third_party/WebKit/LayoutTests/platform/win/fullscreen/model/fully-exit-fullscreen-nested-iframe-expected.txt

CONSOLE WARNING: Upcoming versions will block content-initiated top frame navigations to data: URLs. For more information, see https://goo.gl/BaZAea.
This is a testharness.js-based test.
PASS Fully exit fullscreen with two elements on stack and iframe 
Harness: the test ran to completion.
